services/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pymongo==4.13.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic-settings==2.10.1
email-validator==2.2.0
pydoll-python==2.16.0
